some practical patches

+++ main.py
@@ -14,7 +14,7 @@ TOKEN = os.environ['AUS_BILLS_DISCORD_BOT_TOKEN']
 client = commands.Bot(command_prefix='!')
 
 # vars
-SERVER_ID = 551999201714634752
+server = None
 LOWER_BILLS_CHANNEL_ID = None
 UPPER_BILLS_CHANNEL_ID = None
 ibdd_emojis = ['\u2611', '\u274E', '\U0001F48E', '\U0001F4CA']
@@ -24,12 +24,12 @@ UPPER_HEADER = ',Short Title,Intro Senate,Passed Senate,Intro House,Passed House
 
 LOWER_CHANNEL_NAME = 'lower-house-bills'
 UPPER_CHANNEL_NAME = 'upper-house-bills'
-
+INTRO_HOUSE = "Intro House"
+INTRO_SENATE = "Intro Senate"
 DATA_DIR = 'data'
 LOWER_FILE = "lower.csv"
 UPPER_FILE = "upper.csv"
 url = "https://www.aph.gov.au/Parliamentary_Business/Bills_Legislation/Bills_Lists/Details_page?blsId=legislation%2fbillslst%2fbillslst_c203aa1c-1876-41a8-bc76-1de328bdb726"
-
 
 new_table_lower = None
 old_table_lower = None
@@ -39,18 +39,21 @@ old_table_upper = None
 
 @client.event
 async def on_ready():
+    global server
     print('Bot ready!')
     await data_setup()
     await client.wait_until_ready()
-    if len(client.guilds):
+    if len(client.guilds) == 1:
         server = client.guilds[0]  # gets first in list as there should be only one server
         await discord_server_setup(server)
         await post_new_lower_bill()
         await post_new_upper_bill()
         await remove_completed_lower()
         await remove_completed_upper()
-        await asyncio.sleep(60*4)
+        await asyncio.sleep(60)
         await data_save()
+    elif len(client.guilds) > 1:
+        print("Bot connected to multiple servers \nOnly use one...")
     else:
         print("No server connected...")
     await client.close()
@@ -70,18 +73,22 @@ async def data_setup():
         f = open(DATA_DIR + '/' + UPPER_FILE, 'w')
         f.write(UPPER_HEADER)
         f.close()
-
-    new_table_lower = pd.read_html(url, header=0)[0]
+    try:
+        new_table_lower = pd.read_html(url, header=0)[0]
+        new_table_upper = pd.read_html(url, header=0)[1]
+    except ImportError as e:
+        print('Error: Link may be broken, please check')
     old_table_lower = pd.read_csv(DATA_DIR + '/' + LOWER_FILE)
-    new_table_upper = pd.read_html(url, header=0)[1]
     old_table_upper = pd.read_csv(DATA_DIR + '/' + UPPER_FILE)
 
 
 async def data_save():
     global new_table_lower
     global new_table_upper
-    new_table_lower.to_csv(DATA_DIR + '/' + LOWER_FILE)
-    new_table_upper.to_csv(DATA_DIR + '/' + UPPER_FILE)
+    if (type(pd.DataFrame()) == type(new_table_lower)):
+        new_table_upper.to_csv(DATA_DIR + '/' + LOWER_FILE)
+    if (type(pd.DataFrame()) == type(new_table_upper)):
+        new_table_upper.to_csv(DATA_DIR + '/' + UPPER_FILE)
 
 
 async def discord_server_setup(server):
@@ -114,25 +121,23 @@ async def clear_channel(channel):
     await channel.delete_messages(messages)
 
 
-@client.event
-async def on_message(message):
-    if message.author.id == BOT_ID and (message.channel.id == LOWER_BILLS_CHANNEL_ID or message.channel.id == UPPER_BILLS_CHANNEL_ID):
-        for emoji in ibdd_emojis[:2]:
-            await message.add_reaction(emoji)
+# @client.event
+# async def on_message(message):
+#     if message.author.id == BOT_ID and (message.channel.id == LOWER_BILLS_CHANNEL_ID or message.channel.id == UPPER_BILLS_CHANNEL_ID):
+#         for emoji in ibdd_emojis[:2]:
+#             await message.add_reaction(emoji)
 
 
 async def post_new_upper_bill():
     await client.wait_until_ready()
-    server = client.get_guild(id=SERVER_ID)
     channel = client.get_channel(UPPER_BILLS_CHANNEL_ID)
-    await post_new_bill(channel, old_table_upper, new_table_upper, "Intro Senate")
+    await post_new_bill(channel, old_table_upper, new_table_upper, INTRO_SENATE)
 
 
 async def post_new_lower_bill():
     await client.wait_until_ready()
-    server = client.get_guild(id=SERVER_ID)
     channel = client.get_channel(LOWER_BILLS_CHANNEL_ID)
-    await post_new_bill(channel, old_table_lower, new_table_lower, "Intro House")
+    await post_new_bill(channel, old_table_lower, new_table_lower, INTRO_HOUSE)
 
 
 def check_not_passed(table, row):
@@ -151,48 +156,51 @@ def check_not_passed(table, row):
 
 
 async def post_new_bill(channel, old_table, new_table, date_header):
-    for i in range(len(list(new_table["Short Title"]))):
-        tit = list(new_table["Short Title"])[i]
-        date = list(new_table[date_header])[i]
-        if tit not in list(old_table["Short Title"]) and check_not_passed(new_table, i):
-            print(tit, date)
-            Embed = discord.Embed(title=tit,
-                                  description="Introduced on {}".format(date),
-                                  colour=discord.Colour.purple())
-            Embed.add_field(
-                name="Bill details:", value="[Click here](https://www.aph.gov.au/Parliamentary_Business/Bills_Legislation/Bills_Lists/Details_page?blsId=legislation%2fbillslst%2fbillslst_c203aa1c-1876-41a8-bc76-1de328bdb726)")
-            await channel.send(embed=Embed)
+    if (type(pd.DataFrame()) == type(new_table)):
+        for i in range(len(list(new_table["Short Title"]))):
+            tit = list(new_table["Short Title"])[i]
+            date = list(new_table[date_header])[i]
+            if tit not in list(old_table["Short Title"]) and check_not_passed(new_table, i):
+                print(tit, date)
+                Embed = discord.Embed(title=tit,
+                                      description="Introduced on {}".format(date),
+                                      colour=discord.Colour.purple())
+                Embed.add_field(
+                    name="Bill details:", value="[Click here]({})".format(url))
+                message = await channel.send(embed=Embed)
+                for emoji in ibdd_emojis[:2]:
+                    await message.add_reaction(emoji)
 
 
 async def remove_completed_lower():
     await client.wait_until_ready()
-    server = client.get_guild(id=SERVER_ID)
     channel = client.get_channel(LOWER_BILLS_CHANNEL_ID)
     messages = []
-    async for message in channel.history(limit=100):
-        for i in range(len(list(new_table_lower["Short Title"]))):
-            tit = list(new_table_lower["Short Title"])[i]
+    if (type(pd.DataFrame()) == type(new_table_lower)):
+        async for message in channel.history(limit=100):
+            for i in range(len(list(new_table_lower["Short Title"]))):
+                tit = list(new_table_lower["Short Title"])[i]
 
-            if tit == message.embeds[0].title:
-                if not check_not_passed(new_table_lower, i):
-                    print("Delete:", message.embeds[0].title)
-                    messages.append(message)
-    await channel.delete_messages(messages)
+                if tit == message.embeds[0].title:
+                    if not check_not_passed(new_table_lower, i):
+                        print("Delete:", message.embeds[0].title)
+                        messages.append(message)
+        await channel.delete_messages(messages)
 
 
 async def remove_completed_upper():
     await client.wait_until_ready()
-    server = client.get_guild(id=SERVER_ID)
     channel = client.get_channel(UPPER_BILLS_CHANNEL_ID)
     messages = []
-    async for message in channel.history(limit=100):
-        for i in range(len(list(new_table_upper["Short Title"]))):
-            tit = list(new_table_upper["Short Title"])[i]
-            if tit == message.embeds[0].title:
-                if not check_not_passed(new_table_upper, i):
-                    print("Delete:", message.embeds[0].title)
-                    messages.append(message)
-    await channel.delete_messages(messages)
+    if (type(pd.DataFrame()) == type(new_table_upper)):
+        async for message in channel.history(limit=100):
+            for i in range(len(list(new_table_upper["Short Title"]))):
+                tit = list(new_table_upper["Short Title"])[i]
+                if tit == message.embeds[0].title:
+                    if not check_not_passed(new_table_upper, i):
+                        print("Delete:", message.embeds[0].title)
+                        messages.append(message)
+        await channel.delete_messages(messages)
 
 
 try:
